# dodo payments.openapi.json
{"openapi":"3.0.1","info":{"title":"dodo payments","description":"","version":"1.0.0"},"tags":[],"paths":{"/v1/auth/register":{"post":{"summary":"register a new user","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"email":"benny@gmail.com","password":"randomOne!","full_name":"benny adster","type":"string"},"example":"{\n    \"email\": \"benny@gmail.com\",\n    \"password\": \"randomOne!\",\n    \"full_name\": \"benny adster\",\n}"}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/v1/auth/login":{"post":{"summary":"Login","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"email":"benny@gmail.com","password":"randomOne!"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/v1/auth/refresh":{"post":{"summary":"refresh token","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"refresh-token":"2c0d38dd-752e-457b-b32e-6a34cd4643a2"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/v1/users/uid":{"get":{"summary":"check user","deprecated":false,"description":"","tags":[],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI4OGEzZjQzZi02YzRmLTQ4N2YtYTM0ZC0yOTkwMjU5N2JmOTIiLCJleHAiOjE3MzUwMjM4NjgsImlhdCI6MTczNTAyMjk2OH0.dgqpaZ0FKC3yJNrj_4AxMbMMpffXYpcQLj5OJVAdl9g","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/v1/users/deposit":{"post":{"summary":"deposit","deprecated":false,"description":"","tags":[],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI4OGEzZjQzZi02YzRmLTQ4N2YtYTM0ZC0yOTkwMjU5N2JmOTIiLCJleHAiOjE3MzUwMjkwODMsImlhdCI6MTczNTAyODE4M30.pJfVrm8ENdZC-Uvegr_NYN1kzPOSn8RgmgAvrBFrpW0","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"email":"benny@gmail.com","full_name":"benny adster","amount":"432"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/v1/users/update":{"get":{"summary":"update user","deprecated":false,"description":"","tags":[],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"user_id":"","name":"","email":""}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/v1/tx/transfer":{"post":{"summary":"tx transfer","deprecated":false,"description":"","tags":[],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyZTI3Zjk5OC0yOWIzLTQ4MTgtOTBjOS02ODhiMThiOGM5MTUiLCJleHAiOjE3MzUwMzk5MDUsImlhdCI6MTczNTAzOTAwNX0.ulE1SDmGziJT27lD8z2Oy5X2J8bmIGus0kIWbJOL6nk","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"sender_id":"2e27f998-29b3-4818-90c9-688b18b8c915","receiver_id":"f2a93195-7b35-4aaa-a292-28e262e119c9","amount":"200","description":"personel transfer"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/v1/tx/list_txs":{"get":{"summary":"List txs","deprecated":false,"description":"","tags":[],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJiOGFmM2U5MS0xODkyLTRmYzItYmQyNy00ZjlmNGY0OGE1ODAiLCJleHAiOjE3MzUwOTkzMTksImlhdCI6MTczNTA5ODQxOX0.HeHTiv0DV40jmq2kBvDo3EzJPGvnmDvvBWdauHD4OLw","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/v1/tx/get_tx/{uid}":{"get":{"summary":"get txs","deprecated":false,"description":"","tags":[],"parameters":[{"name":"uid","in":"path","description":"","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"","required":false,"example":"","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}}},"components":{"schemas":{},"securitySchemes":{}},"servers":[]}